refactor(common, footer): improve code
