+++ core/shared-ui/src/commonMain/kotlin/com/neoutils/neoregex/core/sharedui/component/Footer.kt
@@ -49,7 +49,7 @@ sealed class FooterAction {
 }
 
 @Composable
-fun Footer(
+expect fun Footer(
     pattern: TextState,
     history: HistoryState,
     modifier: Modifier = Modifier,
@@ -58,160 +58,4 @@ fun Footer(
     onAction: (FooterAction) -> Unit = {},
     tooling: (@Composable () -> Unit)? = null,
     syntax: Syntax.Regex = remember { Syntax.Regex() },
-) = Column(
-    modifier = modifier,
-) {
-
-    val coroutine = rememberCoroutineScope()
-
-    if (platform.isAndroid) {
-        AnimatedContent(
-            targetState = salvageManager.flow.collectAsStateWithLifecycle().value,
-            transitionSpec = {
-                val showUp = fadeIn() + slideIntoContainer(SlideDirection.Up)
-                val hideDown = fadeOut() + slideOutOfContainer(SlideDirection.Down)
-
-                showUp togetherWith hideDown
-            },
-            contentKey = { it != null }
-        ) { opened ->
-            if (opened != null) {
-                Surface(
-                    shape = RectangleShape,
-                    shadowElevation = dimensions.small,
-                    color = colorScheme.surfaceContainerLow,
-                    contentColor = colorScheme.onSurface,
-                ) {
-                    Salvage(
-                        opened = opened,
-                        textStyle = TextStyle(
-                            fontSize = 16.sp
-                        ),
-                        modifier = Modifier
-                            .padding(dimensions.tiny)
-                            .fillMaxWidth(),
-                        onAction = { action ->
-                            when (action) {
-                                SalvageAction.Close -> {
-                                    coroutine.launch {
-                                        salvageManager.close()
-                                    }
-                                }
-
-                                SalvageAction.Update -> {
-                                    coroutine.launch {
-                                        salvageManager.update()
-                                    }
-                                }
-
-                                is SalvageAction.ChangeName -> {
-                                    coroutine.launch {
-                                        salvageManager.update {
-                                            it.copy(
-                                                title = action.name
-                                            )
-                                        }
-                                    }
-                                }
-
-                                SalvageAction.Reset -> {
-                                    coroutine.launch {
-                                        salvageManager.sync()
-                                    }
-                                }
-                            }
-                        }
-                    )
-                }
-            } else {
-                Spacer(Modifier.fillMaxWidth())
-            }
-        }
-    }
-
-    Surface(
-        shape = RectangleShape,
-        shadowElevation = dimensions.small,
-        color = colorScheme.surfaceContainer,
-        contentColor = colorScheme.onSurface,
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-
-            val textFieldValue = remember(pattern) { pattern.toTextFieldValue() }
-
-            NeoTextField(
-                value = syntax
-                    .highlight
-                    .rememberTextFieldValue(textFieldValue),
-                onValueChange = {
-                    onAction(
-                        FooterAction.UpdateRegex(
-                            it.toText()
-                        )
-                    )
-                },
-                modifier = Modifier
-                    .weight(weight = 1f)
-                    .onFocusChanged(onFocus)
-                    .onPreviewKeyEvent {
-                        when (Command.from(it)) {
-                            Command.UNDO -> {
-                                onAction(
-                                    FooterAction.History.Undo(Field.REGEX)
-                                )
-                                true
-                            }
-
-                            Command.REDO -> {
-                                onAction(
-                                    FooterAction.History.Redo(Field.REGEX)
-                                )
-                                true
-                            }
-
-                            else -> false
-                        }
-                    },
-                singleLine = true,
-                hint = {
-                    Text(
-                        text = stringResource(Res.string.matcher_footer_insert_regex_hint),
-                        style = LocalTextStyle.current.let {
-                            it.copy(
-                                color = it.color.copy(
-                                    alpha = 0.5f
-                                )
-                            )
-                        },
-                        overflow = TextOverflow.Ellipsis,
-                        maxLines = 1
-                    )
-                },
-            )
-
-            tooling?.let {
-                Row(
-                    modifier = Modifier
-                        .padding(vertical = dimensions.small)
-                        .padding(end = dimensions.default)
-                ) {
-                    it()
-                }
-            }
-
-            History(
-                history = history,
-                onAction = onAction,
-                modifier = Modifier
-                    .padding(vertical = dimensions.small)
-                    .padding(end = dimensions.small)
-                    .focusProperties {
-                        canFocus = false
-                    }
-            )
-        }
-    }
-}
+)
